csw/requirements.txt
pygeometa
owslib
psycopg2-binary
python-dotenv
